Upgrade jsoncpp to 1.8.4
This current version supports the new `Json::CharReader`, which
replaces the now-deprecated `Json::Reader`. It's necessary to
eliminate `Json::Reader` from libopenshot because using it
causes a flood of deprecation warnings when building with
system jsoncpp.

Also, the `thirdparty/jsoncpp` directory was generated using
the jsoncpp project's `amalgamate.py` tool, which condenses
the library down to a single C++ source file and two headers,
making it easier to incorporate into libopenshot.